+++ calculator-mvc/CalculatorView.h
@@ -13,7 +13,7 @@ NS_ASSUME_NONNULL_BEGIN
 // 2 methods for prompt, and displaying results
 -(void) welcomeUser;
 -(void) askUserInput: (int) num;
--(void) displayResult: (float) result num1:(int)num1 num2:(int)num2;
+-(void) displayResult: (int) result num1:(int)num1 num2:(int)num2;
 
 @end
